Moved sync package under executer
